\name{print.table.help}
\alias{print.table.help}
\title{Prints myPersonality table information.}
\usage{
  \method{print}{table.help} (x)
}
\arguments{
  \item{x}{An object returned by explainTable}
}
\description{
  This function extends generic print function.
}
\examples{
\dontrun{explainTable("address")}
}
\keyword{attribute}

