\name{explainTable}
\alias{explainTable}
\title{Provide information about specific myPersonality database tables.}
\usage{
  explainTable(table.name)
}
\arguments{
  \item{table.name}{A character string that specifies the
  table that you are interested in.}
}
\description{
  This function retrieves data from the Cambridge
  Psychometrics Centre's myPersonality database.
}
\examples{
# explainTable("address")
}
\keyword{attribute}

